% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{List all combinations of values}
\usage{
cross(..., .type = c("list", "data.frame", "tibble"))
}
\arguments{
\item{...}{Inputs or a \link{list} of inputs}

\item{.type}{Output format.
Either a \verb{"}\code{\link{list}}\verb{"}, \verb{"}\code{\link{data.frame}}\verb{"} or
\verb{"}\code{\link[tibble:tibble]{tibble}}\verb{"}.
Defaults to \code{"list"}.}
}
\value{
A \link{list}, \link{data.frame}, or \link[tibble:tibble]{tibble}, depending on
\code{.type}.
Names will match the names of the inputs.
Unnamed inputs will be left unnamed for lists and data frames, and
automatically named for tibbles.
}
\description{
List all combinations of values
}
\examples{
cross(list(1:3, letters[1:3]))
cross(1:3, letters[1:3])

cross(1:3, letters[1:3], .type = "data.frame")
cross(1:3, letters[1:3], .type = "tibble")

cross(list(number = 1:3, letter = letters[1:3]))
cross(number = 1:3, letter = letters[1:3])
cross(number = 1:3, letter = letters[1:3], .type = "tibble")
}
\seealso{
\code{\link[=cross_join]{cross_join()}} to find combinations of data frame rows.
}
