% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_fit.R
\name{cross_fit}
\alias{cross_fit}
\title{Cross map a model across multiple formulas and multiple subsets}
\usage{
cross_fit(
  data,
  cols,
  formulas,
  fn = stats::lm,
  fn_args = list(),
  tidy = broom::tidy,
  tidy_args = list()
)
}
\arguments{
\item{data}{A data frame}

\item{cols}{Columns to subset the data.
Can be any expression supported by
<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>.}

\item{formulas}{A list of formulas to apply to each subset of the data.
If named, these names will be used in the \code{model} column of the output.
Otherwise, the formulas will be converted to strings in the \code{model} column.}

\item{fn}{The modeling function. Defaults to \code{lm}}

\item{fn_args}{A list of additional arguments to \code{fn}}

\item{tidy}{A logical or a function to use to tidy model output into
data.frame columns.
If \code{FALSE}, no tidying occurs.
Defaults to \code{\link[broom:reexports]{broom::tidy()}}.}

\item{tidy_args}{A list of additional arguments to the \code{tidy} function}
}
\value{
A tibble with subsetting columns,
a column for the model formula applied,
and columns of tidy model output or a list column of models
(if \code{tidy = FALSE})
}
\description{
Applies a modeling function to every combination of a set of formulas and a
set of data subsets.
}
\examples{
cross_fit(mtcars, cyl, mpg ~ wt)
cross_fit(mtcars, cyl, list(mpg ~ wt, mpg ~ hp))
cross_fit(mtcars, cyl, list(wt = mpg ~ wt, hp = mpg ~ hp))

cross_fit(mtcars, c(cyl, vs), list(mpg ~ wt, mpg ~ hp))
cross_fit(mtcars, dplyr::starts_with("c"), list(mpg ~ wt, mpg ~ hp))

cross_fit(
  mtcars, am, list(vs ~ cyl, vs ~ hp),
  fn = glm, fn_args = list(family = binomial(link = logit))
)

cross_fit(mtcars, cyl, list(mpg ~ wt, mpg ~ hp), tidy = FALSE)
cross_fit(mtcars, cyl, list(mpg ~ wt, mpg ~ hp), tidy_args = c(conf.int = TRUE))
}
\seealso{
\code{\link[=xmap]{xmap()}} to apply any function to combinations of inputs
}
