% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_list.R
\name{cross_list}
\alias{cross_list}
\alias{cross_df}
\title{List all combinations of values}
\usage{
cross_list(...)

cross_df(...)
}
\arguments{
\item{...}{Inputs or a \link{list} of inputs}
}
\value{
A \link{list}, \link{data.frame}, or \link[tibble:tibble]{tibble}.
Names will match the names of the inputs.
Unnamed inputs will be left unnamed for \code{cross_list()} and \code{cross_df()},
and automatically named for \code{cross_tbl()}.
}
\description{
List all combinations of values
}
\examples{
fruits   <- c("apple", "banana", "cantaloupe")
desserts <- c("cupcake", "muffin", "streudel")

cross_list(list(fruits, desserts))
cross_list(fruits, desserts)

cross_df(fruits, desserts)
cross_tbl(fruits, desserts)

cross_list(list(fruit = fruits, dessert = desserts))
cross_list(fruit = fruits, dessert = desserts)
cross_df(fruit = fruits, dessert = desserts)
cross_tbl(fruit = fruits, dessert = desserts)
}
\seealso{
\code{\link[=cross_join]{cross_join()}} to find combinations of data frame rows.
}
