% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFALSE.R
\name{isFALSE}
\alias{isFALSE}
\title{Backport of isFALSE() for R < 3.4.2}
\usage{
isFALSE(x)
}
\arguments{
\item{x}{A logical}
}
\value{
\code{TRUE} if \code{x} is \code{FALSE}. \code{FALSE} otherwise.
}
\description{
Backport of isFALSE() for R < 3.4.2
}
