% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossmap-package.R
\docType{package}
\name{crossmap-package}
\alias{crossmap}
\alias{crossmap-package}
\title{crossmap: Apply Functions to All Combinations of List Elements}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides an extension to the 'purrr' family of mapping functions to
    apply a function to each combination of elements in a list of inputs.
    Also includes general purpose functions for finding every combination of
    elements of lists or rows of data frames and a specialized function for
    applying multiple models to multiple subsets of a dataset.
}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
